+++ src/WebJobs.Script/StorageProvider/TableServiceClientProvider.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
-using System.Globalization;
 using Azure.Core;
 using Azure.Data.Tables;
 using Microsoft.Extensions.Azure;
@@ -10,6 +8,12 @@ using Microsoft.Extensions.Configuration;
 
 namespace Microsoft.Azure.WebJobs.Script
 {
+    /// <summary>
+    /// Pass-through class to create a <see cref="TableServiceClient"/> using <see cref="AzureComponentFactory"/>.
+    /// This class uses additional configuration settings for a specified connection in addition to those in <see cref="AzureComponentFactory"/>.
+    /// To support scenarios where a storage connection (i.e. AzureWebJobsStorage) needs to reference multiple serviceUris (blob and queue),
+    /// properties in <see cref="StorageServiceUriOptions"/> are supported to create the Azure table service URI.
+    /// </summary>
     internal sealed class TableServiceClientProvider : StorageClientProvider<TableServiceClient, TableClientOptions>
     {
         /// <summary>
@@ -23,7 +27,7 @@ namespace Microsoft.Azure.WebJobs.Script
 
         protected override TableServiceClient CreateClient(IConfiguration configuration, TokenCredential tokenCredential, TableClientOptions options)
         {
-            // If connection string is present, it will be honored first; bypass creating a serviceUri
+            // Resolve the service URI if the connection string is not defined
             if (!IsConnectionStringPresent(configuration))
             {
                 var serviceUri = configuration.Get<StorageServiceUriOptions>()?.GetTableServiceUri();
@@ -34,53 +38,6 @@ namespace Microsoft.Azure.WebJobs.Script
             }
 
             return base.CreateClient(configuration, tokenCredential, options);
-        }
-
-        /// <summary>
-        /// An options class that constructs a storage service URI from a different properties.
-        /// These properties are specific to WebJobs, as there may be other relevant properties used downstream
-        /// to create storage clients.
-        /// <seealso cref="Microsoft.Extensions.Azure.ClientFactory" />
-        /// A storage service URI can be built using just the account name along with default
-        /// parameters for Scheme and Endpoint Suffix.
-        /// </summary>
-        internal class StorageServiceUriOptions
-        {
-            private const string DefaultScheme = "https";
-            private const string DefaultEndpointSuffix = "core.windows.net";
-
-            /// <summary>
-            /// Gets or sets the resource URI for table storage. If this property is given explicitly, it will be
-            /// honored over the AccountName property.
-            /// </summary>
-            public Uri TableServiceUri { get; set; }
-
-            /// <summary>
-            /// Gets or sets the name of the storage account.
-            /// </summary>
-            public string AccountName { get; set; }
-
-            /// <summary>
-            /// Constructs the table service URI from the properties in this class.
-            /// First checks if TableServiceUri is specified. If not, the AccountName is used
-            /// to construct a table service URI with https scheme and core.windows.net endpoint suffix.
-            /// </summary>
-            /// <returns>Service URI to Azure Table storage.</returns>
-            public Uri GetTableServiceUri()
-            {
-                if (TableServiceUri is not null)
-                {
-                    return TableServiceUri;
-                }
-
-                if (!string.IsNullOrEmpty(AccountName))
-                {
-                    var uri = string.Format(CultureInfo.InvariantCulture, "{0}://{1}.table.{2}", DefaultScheme, AccountName, DefaultEndpointSuffix);
-                    return new Uri(uri);
-                }
-
-                return default;
-            }
         }
     }
 }
